requirements.txt
orjson
rich
dynaconf
pydantic[email]
loguru
pyjwt
fastapi[standard]
hypercorn[h2]
httpx[http2]
httpx-sse
lark
surrealdb==2.0.0-alpha1
dotenv
